Updates prior to bug fixing (Bug in tcl)
tcl8.6.9+ (first release with trampoline code) misinterprets value
of list when used as the args description in ::proc. I can't rely
on this being fixed, or if fixed that users might not update to
a fixed version.
The fix is to change the separator between the element and type
from ":" to "!". Example:
{Email:myservice::Email} changes to {Email!myservice::Email}.