Modified TWiST to work around bug in tcl, see previous commit
Under very rare circumstances Tcl Core 8.6.9+ has a bug which
causes ::proc to fail. The proc command is passed a specification
for the arguments with one or two parameters: name defaultValue.
Tcl sometimes stores the 'bytes' (text value) of a list item
somewhere below the first pointer. This leads Tcl to think that
the argument name has zero length and = 0x0. This leads proc to
announce that the arg's name is "", which it does not like.

I took this opportunity to make slight changes to the processing
of the input info which is a list example:
{Email:xsd::string {minOccurs 0 default \"a@b.c\"}}
this becomes now:
{Email!xsd::string {minOccurs 0 default \"a@b.c\"}}
      ^--- the colon becomes a !. This allows me to use [split]
without much worry that the bang will be part of an element name
or in the type spec.